+++ services/frontend/src/components/reviews/filters/EventFilter.vue
@@ -52,7 +52,7 @@ const filterEvents = (query: string, update: (callback: () => void) => void) => 
     })
   } else {
     api.events
-      .searchEvents({ id_organization: props.organizationId, status: 'PUBLISHED', title: query })
+      .searchEvents({ organizationId: props.organizationId, status: 'PUBLISHED', title: query })
       .then((response) => {
         update(() => {
           hasSearched.value = true