+++ services/interactions/src/events/controllers/participant.consumer.ts
@@ -6,9 +6,20 @@ import { ParticipationService } from '../services/participation.service';
 export class ParticipationConsumer {
   constructor(private readonly participationService: ParticipationService) {}
 
-  // TODO: implement handle events participation
-  @EventPattern('test')
-  handleTicketPurchased(@Payload() data: unknown) {
-    console.log('Received event:', data);
+  @EventPattern('participation.created')
+  async handleParticipationCreated(@Payload() data: unknown) {
+    console.log('Received participation created event:', data);
+    const wrapper = data as {
+      ParticipationCreated: { eventId: string; userId: string };
+    };
+    const payload = wrapper.ParticipationCreated;
+    if (payload && payload.eventId && payload.userId) {
+      await this.participationService.participate(
+        payload.eventId,
+        payload.userId,
+      );
+    } else {
+      console.error('Invalid payload for participation.created event:', data);
+    }
   }
 }