+++ services/interactions/src/metadata/dto/event-published.dto.ts
@@ -1,37 +1,24 @@
 import {
   IsString,
   IsNotEmpty,
-  IsInt,
-  Min,
-  Max,
   IsOptional,
   IsArray,
+  IsDefined,
 } from 'class-validator';
 
-export class CreateReviewDto {
+export class EventPublishedDto {
+  @IsDefined()
   @IsString()
   @IsNotEmpty()
-  userId: string;
+  eventId: string;
 
+  @IsDefined()
   @IsString()
   @IsNotEmpty()
   creatorId: string;
 
+  @IsOptional()
   @IsArray()
   @IsString({ each: true })
-  @IsOptional()
   collaboratorIds?: string[];
-
-  @IsInt()
-  @Min(1)
-  @Max(5)
-  rating: number;
-
-  @IsString()
-  @IsNotEmpty()
-  title: string;
-
-  @IsString()
-  @IsOptional()
-  comment?: string;
 }